core: Move mpi_rescale_particles()
